frontend/streamlit/requirements.txt
streamlit
pika
paho-mqtt
